fix: fix totalQueueSize value, change consume term to 3 sec

See merge request s10-final/S10P31A401!151

+++ src/main/src/main/java/com/qqueueing/main/waiting/model/WaitingStatusDto.java
@@ -15,11 +15,11 @@ public class WaitingStatusDto {
     private Set<String> doneSet = new HashSet<>();
     private List<Long> outList = new LinkedList<>();
     private int lastOffset;
-    private int totalQueueSize;
+    private long totalQueueSize;
     private String targetUrl;
 
     @Builder
-    public WaitingStatusDto(int partitionNo, String targetUrl, int lastOffset, int totalQueueSize) {
+    public WaitingStatusDto(int partitionNo, String targetUrl, int lastOffset, long totalQueueSize) {
         this.partitionNo = partitionNo;
         this.lastOffset = lastOffset;
         this.totalQueueSize = totalQueueSize;
